refactor(tab-manager): inline workspace actions and remove redundant device ID cache
The createTabManagerActions factory was only used in one location and
added a redundant cache layer over getDeviceId() which already caches
internally. Inline the action definitions directly into workspace-popup.ts
and delete the separate workspace-actions.ts file.
